+++ python-package/tests/test_lookup_muni.py
@@ -1,7 +1,7 @@
 import importlib
 
-import pandas as pd
 import pytest
+import pandas as pd
 import geopandas as gpd
 from shapely.geometry import Point
 
@@ -9,36 +9,45 @@ from geobr.lookup_muni import lookup_muni
 
 
 @pytest.fixture
-def mock_lookup_table():
+def mock_seat():
     return gpd.GeoDataFrame(
         {
-            "code_muni": [2304400, 3304557],
-            "name_muni": ["Fortaleza", "Rio de Janeiro"],
-            "abbrev_state": ["CE", "RJ"],
-            "code_state": [23, 33],
+            "code_muni": [3304557, 3550308],
+            "name_muni": ["Rio de Janeiro", "São Paulo"],
+            "abbrev_state": ["RJ", "SP"],
+            "code_state": [33, 35],
         },
-        geometry=[Point(0, 0), Point(1, 1)],
+        geometry=[Point(0, 0)] * 2,
         crs="EPSG:4674",
     )
 
 
-def test_lookup_muni_by_code(mock_lookup_table, monkeypatch):
+def _patch_seat(monkeypatch, mock_seat):
     mod = importlib.import_module("geobr.read_municipal_seat")
-    monkeypatch.setattr(mod, "read_municipal_seat", lambda **k: mock_lookup_table)
-    out = lookup_muni(code_muni=2304400, year=2010)
-    assert isinstance(out, pd.DataFrame)
+    monkeypatch.setattr(mod, "read_municipal_seat", lambda **k: mock_seat)
+
+
+def test_mutual_exclusion(mock_seat, monkeypatch):
+    _patch_seat(monkeypatch, mock_seat)
+    with pytest.raises(ValueError, match="cannot be used"):
+        lookup_muni(name_muni="Rio", code_muni=3304557)
+
+
+def test_lookup_by_code(mock_seat, monkeypatch):
+    _patch_seat(monkeypatch, mock_seat)
+    out = lookup_muni(code_muni=3304557, year=2010)
     assert len(out) == 1
+    assert isinstance(out, pd.DataFrame)
+    assert out.iloc[0]["name_muni"].lower() == "rio de janeiro"
 
 
-def test_lookup_muni_all(mock_lookup_table, monkeypatch):
-    mod = importlib.import_module("geobr.read_municipal_seat")
-    monkeypatch.setattr(mod, "read_municipal_seat", lambda **k: mock_lookup_table)
+def test_lookup_muni_all(mock_seat, monkeypatch):
+    _patch_seat(monkeypatch, mock_seat)
     out = lookup_muni(code_muni="all", year=2010)
-    assert len(out) == len(mock_lookup_table)
+    assert len(out) == len(mock_seat)
 
 
-def test_lookup_muni_mutual_exclusion(mock_lookup_table, monkeypatch):
-    mod = importlib.import_module("geobr.read_municipal_seat")
-    monkeypatch.setattr(mod, "read_municipal_seat", lambda **k: mock_lookup_table)
+def test_lookup_requires_input(mock_seat, monkeypatch):
+    _patch_seat(monkeypatch, mock_seat)
     with pytest.raises(ValueError):
-        lookup_muni(name_muni="Fortaleza", code_muni=2304400, year=2010)
+        lookup_muni(year=2010, name_muni=None, code_muni=None)
